Let the child decide if signals should be fatal
The use of signal-exit in this module was incorrect.

If the parent did not set any handlers on, for example, SIGTERM, then if
it receives a SIGTERM, it would conclude that it should be fatal.  The
signal would be proxied on to the child, and then signal-exit would
remove the handlers from the parent, and re-send the signal to exit
appropriately.  In this case, if the child was listening for that signal
and decided not to exit, then the parent's exit would be inappropriate.

If the parent _had_ assigned listeners to SIGTERM, then signal-exit
would conclude that the signal is _not_ fatal, and do nothing.  That
also means that the signal would not be proxied to the child process at
all!

As of this patch, the parent attaches listeners to all would-be fatal
signals for the current operating system (which signal-exit helpfully
provides for us), but still only exits when and if the child exits.

Note that a SIGKILL will always kill the parent process, _and never the
child process_, because SIGKILL cannot be caught or proxied.  The only
way to do this would be if Node provided a way to truly exec a process
as the new foreground program in the same process space, without forking
a separate child process.