refactor: add `proxyStdio` internal function
# Why

`foregroundChild` is currently hard-coded to only support forwarding with the root process. To support a general `proxy` function between any two objects behaving like processes, we can begin by refactoring the internals to be more parametrized.

# What

Add a function to start forwarding standard IO. This function is unused at the moment because `foregroundChild` defaults to using the `stdio: "inherit"` `spawn` option, causing the std streams to be `null` on the child process.