BREAKING CHANGE: require Node 8
## Why

Node 6 is no longer supported by the test framework used for this library, meaning that Node 6 can no longer be checked by CI. Also, Node 6 is no longer reached end of life and is no longer maintained. Because of this, the new Node version required should be updated to Node 8.

## What

- Update CI to run against the currently supported Node versions: Node 8, 10 and 12.
- Update `engines.node` in `package.json`.